requirements.txt
appdirs==1.4.3
attrs==19.3.0
cachetools==4.2.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dataclasses==0.7
distlib==0.3.0
env==0.1.0
fastapi==0.61.2
filelock==3.0.12
Flask==1.1.2
freeze==3.0
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-datastore==2.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
h11==0.11.0
idna==2.10
importlib-metadata==1.6.0
importlib-resources==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.16
MarkupSafe==1.1.1
more-itertools==8.2.0
mypy-extensions==0.4.3
numpy==1.19.1
packaging==20.3
pipenv==2018.11.26
pluggy==0.13.1
proto-plus==1.13.0
protobuf==3.14.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.2
pyparsing==2.4.7
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.14.0
SQLAlchemy==1.3.19
starlette==0.13.6
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
uvicorn==0.12.2
virtualenv==20.0.18
virtualenv-clone==0.5.4
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
